issue-13: Refactor player.py from entities
Still a lot of code to clean up, but moving it out.
