+++ python/game/entities/__init__.py
@@ -1,0 +1,7 @@
+""""
+Game entities module
+"""
+from .player import Player
+from .entities import Obstacle, GameState
+
+all__ = ['Player', 'Obstacle', 'GameState']
